ai_detector/requirements.txt
# Core ML dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0

# Training and evaluation
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.7.0
tqdm>=4.62.0

# Data handling
pandas>=1.3.0

# Supabase integration (optional)
supabase>=1.0.0

# Utilities
json5